[Feature] Redirects state on game end
