+++ public/app/layout/battlefield/battlefield.controller.js
@@ -23,12 +23,6 @@
       return bf.get(person + 'Health');
     };
 
-    vm.getRoundWinner = function() {
-      return bf.get('roundWinner');
-    };
-
-    vm.getSubmitted = function() {
-      return bf.get('submitted');
-    };
+    vm.get = bf.get;
   }
 })();
