+++ server/game/listeners.js
@@ -1,7 +1,17 @@
-var choice = function(socket, player) {
+var onDisconnect = function(player) {
+  console.log(player + ' socket disconnected!');
+  this.terminate('a player disconnected.');
+};
 
-  socket.on('choice', function(data) {
-    console.log('player ' + player.id + ' submitted a choice.');
-    player.updateChoice(data.choice);
-  });
+var onChoice = function(player, data) {
+  console.log('inside of onChoice ====== ');
+  console.log('player = ', player);
+  console.log('inside of onChoice. data.choice = ', data.choice);
+  console.log('========================= ');
+  this[player].updateChoice(data.choice);
+};
+
+module.exports = {
+  onDisconnect: onDisconnect,
+  onChoice: onChoice
 };
